+++ tango-pvp/src/game/bn4.rs
@@ -401,7 +401,7 @@ impl crate::hooks::Hooks for Hooks {
                 let match_ = match_.clone();
                 let munger = self.munger();
                 Box::new(move |mut core| {
-                    let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                    let pc = core.as_ref().gba().cpu().thumb_pc();
                     core.gba_mut().cpu_mut().set_thumb_pc(pc + 4);
                     munger.set_copy_data_input_state(core, if match_.blocking_lock().is_some() { 2 } else { 4 });
                 })
@@ -655,7 +655,7 @@ impl crate::hooks::Hooks for Hooks {
             (self.offsets.rom.in_battle_call_handle_link_cable_input, {
                 let munger = self.munger();
                 Box::new(move |mut core| {
-                    let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                    let pc = core.as_ref().gba().cpu().thumb_pc();
                     core.gba_mut().cpu_mut().set_thumb_pc(pc + 4);
                     munger.set_copy_data_input_state(core, 2);
                 })
@@ -897,7 +897,7 @@ impl crate::hooks::Hooks for Hooks {
                     if !stepper_state.disable_bgm() {
                         return;
                     }
-                    let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                    let pc = core.as_ref().gba().cpu().thumb_pc();
                     core.gba_mut().cpu_mut().set_thumb_pc(pc + 4);
                 })
             }),
@@ -922,7 +922,7 @@ impl crate::hooks::Hooks for Hooks {
             (self.offsets.rom.in_battle_call_handle_link_cable_input, {
                 let munger = self.munger();
                 Box::new(move |mut core| {
-                    let pc = core.as_ref().gba().cpu().thumb_pc() as u32;
+                    let pc = core.as_ref().gba().cpu().thumb_pc();
                     core.gba_mut().cpu_mut().set_thumb_pc(pc + 4);
                     munger.set_copy_data_input_state(core, 2);
                 })
